Merge Side branch into main

+++ code/lesson/diadic-image-processing/python.py
@@ -1,7 +1,7 @@
 import machinevisiontoolbox as mv
 
 # 5:24
-cam = mv.VideoFile("traffic_sequence.mp4",dtype = "float32", grey = 1)
+cam = mv.VideoFile("traffic_sequence.mp4",dtype = "float64", grey = 1)
 images = (im for im in cam)
 
 # 5:44
@@ -12,4 +12,3 @@ im2 = next(images, None)
 
 # 6:02
 mv.idisp(im1.image - im2.image)
-
